% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplotNull}
\alias{barplotNull}
\title{Percentage of null counts per sample}
\usage{
barplotNull(counts, group, col = c("lightblue", "orange", "MediumVioletRed",
  "SpringGreen"))
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the bars (one color per biological condition)}
}
\value{
A file named barplotNull.png in the figures directory
}
\description{
Bar plot of the percentage of null counts per sample
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

