% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplotTotal}
\alias{barplotTotal}
\title{Total number of reads per sample}
\usage{
barplotTotal(counts, group, col = c("lightblue", "orange", "MediumVioletRed",
  "SpringGreen"))
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the bars (one color per biological condition)}
}
\value{
A file named barplotTotal.png in the figures directory
}
\description{
Bar plot of the total number of reads per sample
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

