% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCVPlot.R
\name{BCVPlot}
\alias{BCVPlot}
\title{BCV plot (for edgeR dispersions)}
\usage{
BCVPlot(dge, outfile = TRUE)
}
\arguments{
\item{dge}{a \code{DGEList} object}

\item{outfile}{TRUE to export the figure in a png file}
}
\value{
A file named BCV.png in the figures directory with a BCV plot produced by the \code{plotBCV()} function of the edgeR package
}
\description{
Biological Coefficient of Variation plot (for edgeR objects)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

