% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairoSizeWrapper.R
\name{cairoSizeWrapper}
\alias{cairoSizeWrapper}
\title{Cairo Size Wrapper}
\usage{
cairoSizeWrapper(pixel_re)
}
\arguments{
\item{pixel_re}{requested pixel size}
}
\value{
the minimum between the requested and maximal allowed pixel size if Cairo is selected otherwise the requested pixel size
}
\description{
Limits the pixel size to either to the minimum of the given and the max allowed pixel size 32767x32767 pixels if
Cairo graphical has been choosen
}
\author{
vipul patel
}
