% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairoSizeWrapper.R
\name{carioSizeWrapper}
\alias{carioSizeWrapper}
\title{Cario Size Wrapper}
\usage{
carioSizeWrapper(pixel_re)
}
\arguments{
\item{requested}{pixel size}
}
\value{
the minimum between the requested and maximal allowed pixel size if cario is selected otherwise the requested pixel size
}
\description{
Limits the pixel size to either to the minimum of the given and the max allowed pixel size 32767x32767 pixels if
cario graphical has been choosen
}
\author{
vipul patel
}
