% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkParameters.DESeq2.r
\name{checkParameters.DESeq2}
\alias{checkParameters.DESeq2}
\title{Check the parameters (when using DESeq2)}
\usage{
checkParameters.DESeq2(projectName, author, targetFile, rawDir,
  featuresToRemove, varInt, condRef, batch, fitType, cooksCutoff,
  independentFiltering, alpha, pAdjustMethod, typeTrans, locfunc, colors)
}
\arguments{
\item{projectName}{name of the project}

\item{author}{author of the statistical analysis/report}

\item{targetFile}{path to the design/target file}

\item{rawDir}{path to the directory containing raw counts files}

\item{featuresToRemove}{names of the features to be removed}

\item{varInt}{factor of interest}

\item{condRef}{reference biological condition}

\item{batch}{blocking factor in the design}

\item{fitType}{mean-variance relationship: "parametric" (default) or "local"}

\item{cooksCutoff}{outliers detection (TRUE or FALSE)}

\item{independentFiltering}{TRUE/FALSE to perform independent filtering}

\item{alpha}{threshold of statistical significance}

\item{pAdjustMethod}{p-value adjustment method: "BH" (default) or "BY" for example}

\item{typeTrans}{transformation for PCA/clustering: "VST" ou "rlog"}

\item{locfunc}{"median" (default) or "shorth" to estimate the size factors}

\item{colors}{vector of colors of each biological condition on the plots}
}
\value{
A boolean indicating if there is a problem in the parameters
}
\description{
Check the format and the validity of the parameters which will be used for the analysis with DESeq2.
}
\author{
Hugo Varet
}

