% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParameters.edgeR.r
\name{checkParameters.edgeR}
\alias{checkParameters.edgeR}
\title{Check the parameters (when using edgeR)}
\usage{
checkParameters.edgeR(projectName, author, targetFile, rawDir, featuresToRemove,
  varInt, condRef, batch, alpha, pAdjustMethod, cpmCutoff, gene.selection,
  normalizationMethod, colors)
}
\arguments{
\item{projectName}{name of the project}

\item{author}{author of the statistical analysis/report}

\item{targetFile}{path to the design/target file}

\item{rawDir}{path to the directory containing raw counts files}

\item{featuresToRemove}{names of the features to be removed}

\item{varInt}{factor of interest}

\item{condRef}{reference biological condition}

\item{batch}{blocking factor in the design}

\item{alpha}{threshold of statistical significance}

\item{pAdjustMethod}{p-value adjustment method: \code{"BH"} (default) or \code{"BY"} for example}

\item{cpmCutoff}{counts-per-million cut-off to filter low counts}

\item{gene.selection}{selection of the features in MDSPlot}

\item{normalizationMethod}{normalization method: \code{"TMM"} (default), \code{"RLE"} (DESeq) or \code{"upperquartile"}}

\item{colors}{vector of colors of each biological condition on the plots}
}
\value{
A boolean indicating if there is a problem in the parameters
}
\description{
Check the format and the validity of the parameters which will be used for the analysis with edgeR.
}
\author{
Hugo Varet
}
