% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlot.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Clustering of the samples}
\usage{
clusterPlot(counts.trans, group, outfile = TRUE)
}
\arguments{
\item{counts.trans}{a matrix a transformed counts (VST- or rlog-counts if use of DESeq2 or cpm-counts if use of edgeR)}

\item{group}{factor vector of the condition from which each sample belongs}

\item{outfile}{TRUE to export the figure in a png file}
}
\value{
A file named cluster.png in the figures directory with the dendrogram of the clustering
}
\description{
Clustering of the samples based on VST- or rlog-counts (if use of DESeq2) or cpm-counts (if use of edgeR)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
