% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countsBoxplots.R
\name{countsBoxplots}
\alias{countsBoxplots}
\title{Box-plots of (normalized) counts distribution per sample}
\usage{
countsBoxplots(object, group, col = c("lightblue", "orange",
  "MediumVioletRed", "SpringGreen"))
}
\arguments{
\item{object}{a \code{DESeqDataSet} object from DESeq2 or a \code{DGEList} object from edgeR}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the boxplots (one per biological condition)}
}
\value{
A file named countsBoxplots.png in the figures directory containing boxplots of the raw and normalized counts
}
\description{
Box-plots of raw and normalized counts distributions per sample to assess the effect of the normalization
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

