% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportResults.DESeq2.R
\name{exportResults.DESeq2}
\alias{exportResults.DESeq2}
\title{Export results for DESeq2 analyses}
\usage{
exportResults.DESeq2(out.DESeq2, group, alpha = 0.05)
}
\arguments{
\item{out.DESeq2}{the result of \code{run.DESeq2()}}

\item{group}{factor vector of the condition from which each sample belongs}

\item{alpha}{threshold to apply to adjusted p-values}
}
\value{
A list of \code{data.frame} containing counts, pvalues, FDR, log2FC...
}
\description{
Export counts and DESeq2 results
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

