% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.edgeR.R
\name{exportResults.edgeR}
\alias{exportResults.edgeR}
\title{Export results for edgeR analyses}
\usage{
exportResults.edgeR(out.edgeR, group, counts, alpha = 0.05,
  export = TRUE)
}
\arguments{
\item{out.edgeR}{the result of \code{run.edgeR()}}

\item{group}{factor vector of the condition from which each sample belongs}

\item{counts}{non-filtered counts (used to keep them in the final table)}

\item{alpha}{threshold to apply to adjusted p-values}

\item{export}{TRUE to export the results (up, down and complete table) in the tables directory}
}
\value{
A list of \code{data.frame} containing counts, pvalues, FDR, log2FC...
}
\description{
Export counts and edgeR results
}
\details{
\code{counts} are used as input just in order to export features with null counts too.
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
