% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCountData.R
\name{loadCountData}
\alias{loadCountData}
\title{Load count files}
\usage{
loadCountData(
  target,
  rawDir = "raw",
  skip = 0,
  featuresToRemove = c("alignment_not_unique", "ambiguous", "no_feature", "not_aligned",
    "too_low_aQual")
)
}
\arguments{
\item{target}{target \code{data.frame} of the project returned by \code{loadTargetFile()}}

\item{rawDir}{path to the directory containing the count files}

\item{skip}{number of lines of the data file to skip before beginning to read data}

\item{featuresToRemove}{vector of feature Ids (or character string common to feature Ids) to remove from the counts}
}
\value{
The \code{matrix} of raw counts with row names corresponding to the feature Ids and column names to the sample names as provided in the first column of the target.
}
\description{
Load one count file per sample thanks to the file names in the target file.
}
\details{
If \code{featuresToRemove} is equal to \code{"rRNA"}, all the features containing the character string "rRNA" will be removed from the counts.
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
