% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadTargetFile.R
\name{loadTargetFile}
\alias{loadTargetFile}
\title{Load target file}
\usage{
loadTargetFile(targetFile, varInt, condRef, batch)
}
\arguments{
\item{targetFile}{path to the target file}

\item{varInt}{variable on which sorting the target}

\item{condRef}{reference condition of \code{varInt}}

\item{batch}{batch effect to take into account}
}
\value{
A \code{data.frame} containing the informations about the samples (name, file containing the counts and biological condition)
}
\description{
Load the target file containing sample information
}
\details{
The \code{batch} parameter is used only to check if it is available in the target file before running the suite of the script.
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

