% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/majSequences.R
\name{majSequences}
\alias{majSequences}
\title{Most expressed sequences per sample}
\usage{
majSequences(counts, n = 3, group, col = c("lightblue", "orange",
  "MediumVioletRed", "SpringGreen"), outfile = TRUE)
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{n}{number of most expressed sequences to return}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the bars (one per biological condition)}

\item{outfile}{TRUE to export the figure in a png file}
}
\value{
A \code{matrix} with the percentage of reads of the three most expressed sequences and a file named majSeq.png in the figures directory
}
\description{
Proportion of reads associated with the three most expressed sequences per sample
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

