% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDiffTotal.r
\name{nDiffTotal}
\alias{nDiffTotal}
\title{Number of differentially expressed features per comparison}
\usage{
nDiffTotal(complete, alpha = 0.05)
}
\arguments{
\item{complete}{list of \code{data.frame} containing features results (from \code{exportResults.DESeq2()} or \code{exportResults.edgeR()})}

\item{alpha}{threshold to apply to the FDR}
}
\value{
A matrix with the number of up, down and total of features per comparison
}
\description{
Number of down- and up-regulated features per comparison
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
