% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlot.R
\name{PCAPlot}
\alias{PCAPlot}
\title{PCA of samples (if use of DESeq2)}
\usage{
PCAPlot(counts.trans, group, n = min(500, nrow(counts.trans)),
  col = c("lightblue", "orange", "MediumVioletRed", "SpringGreen"),
  outfile = TRUE)
}
\arguments{
\item{counts.trans}{a matrix a transformed counts (VST- or rlog-counts)}

\item{group}{factor vector of the condition from which each sample belongs}

\item{n}{number of features to keep among the most variant}

\item{col}{colors to use (one per biological condition)}

\item{outfile}{TRUE to export the figure in a png file}
}
\value{
A file named PCA.png in the figures directory with a pairwise plot of the three first principal components
}
\description{
Principal Component Analysis of samples based on the 500 most variant features on VST- or rlog-counts (if use of DESeq2)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

