% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeNull.R
\name{removeNull}
\alias{removeNull}
\title{Remove features with null counts in all samples}
\usage{
removeNull(counts)
}
\arguments{
\item{counts}{\code{matrix} of raw counts}
}
\value{
The \code{matrix} of counts without features with only null counts
}
\description{
Remove features with null counts in all samples. These features do not contain any information and will not be used for the statistical analysis.
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

