% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run.DESeq2}
\alias{run.DESeq2}
\title{Wrapper to run DESeq2}
\usage{
run.DESeq2(counts, target, varInt, batch = NULL, locfunc = "median",
  fitType = "parametric", pAdjustMethod = "BH", cooksCutoff = NULL,
  independentFiltering = TRUE, alpha = 0.05, ...)
}
\arguments{
\item{counts}{\code{matrix} of raw counts}

\item{target}{target \code{data.frame} of the project}

\item{varInt}{name of the factor of interest (biological condition)}

\item{batch}{batch effect to take into account (\code{NULL} by default)}

\item{locfunc}{\code{"median"} (default) or \code{"shorth"} to estimate the size factors}

\item{fitType}{mean-variance relationship: "parametric" (default) or "local"}

\item{pAdjustMethod}{p-value adjustment method: \code{"BH"} (default) or \code{"BY"} for instance}

\item{cooksCutoff}{outliers detection threshold (\code{NULL} to let DESeq2 choosing it)}

\item{independentFiltering}{\code{TRUE} or \code{FALSE} to perform the independent filtering or not}

\item{alpha}{significance threshold to apply to the adjusted p-values}

\item{...}{optional arguments to be passed to \code{nbinomWaldTest()}}
}
\value{
A list containing the \code{dds} object (\code{DESeqDataSet} class), the \code{results} objects (\code{DESeqResults} class) and the vector of size factors
}
\description{
Wrapper to run DESeq2: create the \code{DESeqDataSet}, normalize data, estimate dispersions, statistical testing...
}
\author{
Hugo Varet
}

