% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run.edgeR}
\alias{run.edgeR}
\title{Wrapper to run edgeR}
\usage{
run.edgeR(counts, target, varInt, condRef, batch = NULL, cpmCutoff = 1,
  pAdjustMethod = "BH", ...)
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{target}{target \code{data.frame} of the project}

\item{varInt}{name of the factor of interest (biological condition)}

\item{condRef}{reference biological condition}

\item{batch}{batch effect to take into account (\code{NULL} by default)}

\item{cpmCutoff}{counts-per-million cut-off to filter low counts}

\item{pAdjustMethod}{p-value adjustment method: \code{"BH"} (default) or \code{"BY"}}

\item{...}{optional arguments to be passed to \code{glmFit()}}
}
\value{
A list containing the \code{dge} object and the \code{results} object
}
\description{
Wrapper to run edgeR: create the \code{dge} object, normalize data, estimate dispersions, statistical testing...
}
\author{
Hugo Varet
}

