% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARTools-package.r
\docType{package}
\name{SARTools-package}
\alias{SARTools-package}
\title{Statistical Analysis of RNA-Seq Tools}
\description{
SARTools provides R tools and an environment for the statistical analysis of RNA-Seq projects: load and clean data, produce figures, perform statistical analysis/testing with DESeq2 or edgeR, export results and create final report
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

