% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERE.r
\name{SERE}
\alias{SERE}
\title{Pairwise SERE for two samples}
\usage{
SERE(observed)
}
\arguments{
\item{observed}{\code{matrix} with two columns containing observed counts of two samples}
}
\value{
The SERE coefficient for the two samples
}
\description{
Compute the SERE coefficient for two samples
}
\author{
See paper published
}
\references{
Schulze, Kanwar, Golzenleuchter et al, SERE: Single-parameter quality control and sample comparison for RNA-Seq, BMC Genomics, 2012
}

