% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizeResults.DESeq2.r
\name{summarizeResults.DESeq2}
\alias{summarizeResults.DESeq2}
\title{Summarize DESeq2 analysis}
\usage{
summarizeResults.DESeq2(out.DESeq2, group, independentFiltering = TRUE,
  cooksCutoff = NULL, alpha = 0.05, col = c("lightblue", "orange",
  "MediumVioletRed", "SpringGreen"))
}
\arguments{
\item{out.DESeq2}{the result of \code{run.DESeq2()}}

\item{group}{factor vector of the condition from which each sample belongs}

\item{independentFiltering}{\code{TRUE} or \code{FALSE} to perform the independent filtering or not}

\item{cooksCutoff}{Cook's distance threshold for detecting outliers (\code{Inf}
to disable the detection, \code{NULL} to keep DESeq2 threshold)}

\item{alpha}{significance threshold to apply to the adjusted p-values}

\item{col}{colors for the plots}
}
\value{
A list containing: (i) a list of \code{data.frames} from \code{exportResults.DESeq2()}, (ii) the table summarizing the independent filtering procedure and (iii) a table summarizing the number of differentially expressed features
}
\description{
Summarize DESeq2 analysis: diagnotic plots, dispersions plot, summary of the independent filtering, export results...
}
\author{
Hugo Varet
}

