% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tabIndepFiltering.R
\name{tabIndepFiltering}
\alias{tabIndepFiltering}
\title{Table of the number of features discarded by the independent filtering (if use of DESeq2)}
\usage{
tabIndepFiltering(results)
}
\arguments{
\item{results}{list of results of \code{results(dds,...)} with chosen parameters}
}
\value{
A \code{matrix} with the threshold and the number of features discarded for each comparison
}
\description{
Compute the number of features discarded by the independent filtering for each comparison (if use of DESeq2)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

