% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.r
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plots}
\usage{
volcanoPlot(complete, alpha = 0.05, outfile = TRUE, padjlim = NULL)
}
\arguments{
\item{complete}{A \code{list} of \code{data.frame} containing features results (from \code{exportResults.DESeq2()} or \code{exportResults.edgeR()})}

\item{alpha}{cut-off to apply on each adjusted p-value}

\item{outfile}{TRUE to export the figure in a png file}

\item{padjlim}{numeric value between 0 and 1 for the adjusted p-value upper limits for all the volcano plots produced (NULL by default to set them automatically)}
}
\value{
A file named volcanoPlot.png in the figures directory containing one volcano plot per comparison
}
\description{
Volcano plot for each comparison: -log10(adjusted P value) vs log2(FC) with one dot per feature (red dot for a differentially expressed feature, black dot otherwise)
}
\author{
Hugo Varet
}
