% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.DESeq2.r
\name{writeReport.DESeq2}
\alias{writeReport.DESeq2}
\title{Write HTML report for DESeq2 analyses}
\usage{
writeReport.DESeq2(target, counts, out.DESeq2, summaryResults,
  majSequences, workDir, projectName, author, targetFile, rawDir,
  featuresToRemove, varInt, condRef, batch, fitType, cooksCutoff,
  independentFiltering, alpha, pAdjustMethod, typeTrans, locfunc, colors)
}
\arguments{
\item{target}{target \code{data.frame} of the project returned by \code{loadTargetFile()}}

\item{counts}{\code{matrix} of counts returned by \code{loadCountData()}}

\item{out.DESeq2}{the result of \code{run.DESeq2()}}

\item{summaryResults}{the result of \code{summarizeResults.DESeq2()}}

\item{majSequences}{the result of \code{descriptionPlots()}}

\item{workDir}{working directory}

\item{projectName}{name of the project}

\item{author}{name of the author of the analysis}

\item{targetFile}{path to the target file}

\item{rawDir}{path to the directory containing the counts files}

\item{featuresToRemove}{vector of features to remove from the counts matrix}

\item{varInt}{factor of interest (biological condition)}

\item{condRef}{reference condition for the factor of interest}

\item{batch}{variable to take as a batch effect}

\item{fitType}{mean-variance relationship: \code{"parametric"} (default) or \code{"local"}}

\item{cooksCutoff}{outliers detection threshold}

\item{independentFiltering}{\code{TRUE} or \code{FALSE} to perform the independent filtering or not}

\item{alpha}{threshold of statistical significance}

\item{pAdjustMethod}{p-value adjustment method: \code{"BH"} or \code{"BY"} for instance}

\item{typeTrans}{transformation for PCA/clustering: \code{"VST"} or \code{"rlog"}}

\item{locfunc}{\code{"median"} (default) or \code{"shorth"} to estimate the size factors}

\item{colors}{vector of colors of each biological condition on the plots}
}
\description{
Write HTML report from graphs and tables created during the analysis with DESeq2
}
\details{
This function generates the HTML report for a statistical analysis with DESeq2. It uses the tables and graphs created during the workflow as well as the parameters defined at the beginning of the script.
}
\author{
Hugo Varet
}
