% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeReport.edgeR.r
\name{writeReport.edgeR}
\alias{writeReport.edgeR}
\title{Write HTML report for edgeR analyses}
\usage{
writeReport.edgeR(target, counts, out.edgeR, summaryResults, majSequences,
  workDir, projectName, author, targetFile, rawDir, featuresToRemove, varInt,
  condRef, batch, alpha, pAdjustMethod, colors, gene.selection,
  normalizationMethod)
}
\arguments{
\item{target}{target \code{data.frame} of the project returned by \code{loadTargetFile()}}

\item{counts}{\code{matrix} of counts returned by \code{loadCountData()}}

\item{out.edgeR}{the result of \code{run.edgeR()}}

\item{summaryResults}{the result of \code{summarizeResults.DESeq2()}}

\item{majSequences}{the result of \code{descriptionPlots()}}

\item{workDir}{path to the working directory}

\item{projectName}{name of the project}

\item{author}{name of the author of the analysis}

\item{targetFile}{path to the target file}

\item{rawDir}{path to the directory containing the counts files}

\item{featuresToRemove}{vector of features to remove from the counts matrix}

\item{varInt}{factor of interest (biological condition)}

\item{condRef}{reference condition for the factor of interest}

\item{batch}{variable to take as a batch effect}

\item{alpha}{threshold of statistical significance}

\item{pAdjustMethod}{p-value adjustment method: \code{"BH"} (default) or \code{"BY"}}

\item{colors}{vector of colors of each biological condition on the plots}

\item{gene.selection}{selection of the features in \code{MDSPlot()} (\code{"pairwise"} by default)}

\item{normalizationMethod}{normalization method: \code{"TMM"} (default), \code{"RLE"} (DESeq) or \code{"upperquartile"}}
}
\description{
Write HTML report from graphs and tables created during the analysis with edgeR
}
\details{
This function generates the HTML report for a statistical analysis with edgeR. It uses the tables and graphs created during the workflow as well as the parameters defined at the beginning of the script.
}
\author{
Hugo Varet
}

