% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotNull.R
\name{barplotNull}
\alias{barplotNull}
\title{Percentage of null counts per sample}
\usage{
barplotNull(
  counts,
  group,
  col = c("lightblue", "orange", "MediumVioletRed", "SpringGreen"),
  outfile = TRUE,
  ggplot_theme = theme_gray()
)
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the bars (one color per biological condition)}

\item{outfile}{TRUE to export the figure in a png file}

\item{ggplot_theme}{ggplot2 theme function (\code{theme_gray()} by default)}
}
\value{
A file named barplotNull.png in the figures directory
}
\description{
Bar plot of the percentage of null counts per sample
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
