% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BCVPlot.R
\name{BCVPlot}
\alias{BCVPlot}
\title{BCV plot (for edgeR dispersions)}
\usage{
BCVPlot(dge)
}
\arguments{
\item{dge}{a \code{DGEList} object}
}
\value{
A file named BCV.png in the figures directory with a BCV plot produced by the \code{plotBCV()} function of the edgeR package
}
\description{
Biological Coefficient of Variation plot (for edgeR objects)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

