% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clusterPlot}
\alias{clusterPlot}
\title{Clustering of the samples}
\usage{
clusterPlot(counts.trans, group)
}
\arguments{
\item{counts.trans}{a matrix a transformed counts (VST- or rlog-counts if use of DESeq2 or cpm-counts if use of edgeR)}

\item{group}{factor vector of the condition from which each sample belongs}
}
\value{
A file named cluster.png in the figures directory with the dendrogram of the clustering
}
\description{
Clustering of the samples based on VST- or rlog-counts (if use of DESeq2) or cpm-counts (if use of edgeR)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

