% Generated by roxygen2 (4.0.2): do not edit by hand
\name{densityPlot}
\alias{densityPlot}
\title{Density plot of all samples}
\usage{
densityPlot(counts, group, col = c("lightblue", "orange", "MediumVioletRed",
  "SpringGreen"))
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors of the curves (one per biological condition)}
}
\value{
A file named densplot.png in the figures directory
}
\description{
Estimation the counts density for each sample
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

