% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionPlots.r
\name{descriptionPlots}
\alias{descriptionPlots}
\title{Description plots of the counts}
\usage{
descriptionPlots(counts, group, col = c("lightblue", "orange",
  "MediumVioletRed", "SpringGreen"))
}
\arguments{
\item{counts}{\code{matrix} of counts}

\item{group}{factor vector of the condition from which each sample belongs}

\item{col}{colors for the plots (one per biological condition)}
}
\value{
PNG files in the "figures" directory and the matrix of the most expressed sequences
}
\description{
Description plots of the counts according to the biological condition
}
\author{
Hugo Varet
}

