% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diagSizeFactorsPlots.R
\name{diagSizeFactorsPlots}
\alias{diagSizeFactorsPlots}
\title{Assess the estimations of the size factors}
\usage{
diagSizeFactorsPlots(dds)
}
\arguments{
\item{dds}{a \code{DESeqDataSet} object}
}
\value{
Two files in the figures directory: diagSizeFactorsHist.png containing one histogram per sample and diagSizeFactorsTC.png for a plot of the size factors vs the total number of reads
}
\description{
Plots to assess the estimations of the size factors
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

