% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dispersionsPlot}
\alias{dispersionsPlot}
\title{Plots about DESeq2 dispersions}
\usage{
dispersionsPlot(dds)
}
\arguments{
\item{dds}{a \code{DESeqDataSet} object}
}
\value{
A file named dispersionsPlot.png in the figures directory containing the plot of the mean-dispersion relationship and a diagnostic of log normality of the dispersions
}
\description{
A plot of the mean-dispersion relationship and a diagnostic of log normality of the dispersions (if use of DESeq2)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

