% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportResults.DESeq2}
\alias{exportResults.DESeq2}
\title{Export results for DESeq2 analyses}
\usage{
exportResults.DESeq2(out.DESeq2, group, cooksCutoff = NULL, alpha = 0.05)
}
\arguments{
\item{out.DESeq2}{the result of \code{run.DESeq2()}}

\item{group}{factor vector of the condition from which each sample belongs}

\item{cooksCutoff}{Cook's distance threshold for detecting outliers (\code{Inf}
to disable the detection, \code{NULL} to keep DESeq2 threshold)}

\item{alpha}{threshold to apply to adjusted p-values}
}
\value{
A list of \code{data.frame} containing counts, pvalues, FDR, log2FC...
}
\description{
Export counts and DESeq2 results
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

