% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAPlot.R
\name{MAPlot}
\alias{MAPlot}
\title{MA-plots}
\usage{
MAPlot(complete, alpha = 0.05, outfile = TRUE)
}
\arguments{
\item{complete}{A \code{list} of \code{data.frame} containing features results (from \code{exportResults.DESeq2()} or \code{exportResults.edgeR()})}

\item{alpha}{cut-off to apply on each adjusted p-value}

\item{outfile}{TRUE to export the figure in a png file}
}
\value{
A file named MAPlot.png in the figures directory containing one MA-plot per comparison
}
\description{
MA-plot for each comparison: log2(FC) vs mean of normalized counts with one dot per feature (red dot for a differentially expressed feature, black dot otherwise)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

