% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSPlot.R
\name{MDSPlot}
\alias{MDSPlot}
\title{MDS plot (for edgeR objects)}
\usage{
MDSPlot(
  dge,
  group,
  n = min(500, nrow(dge$counts)),
  gene.selection = c("pairwise", "common"),
  col = c("lightblue", "orange", "MediumVioletRed", "SpringGreen"),
  outfile = TRUE,
  ggplot_theme = theme_gray()
)
}
\arguments{
\item{dge}{a \code{DGEList} object}

\item{group}{vector of the condition from which each sample belongs}

\item{n}{number of features to keep among the most variant}

\item{gene.selection}{\code{"pairwise"} to choose the top features separately for each pairwise comparison between the samples or \code{"common"} to select the same features for all comparisons. Only used when \code{method="logFC"}}

\item{col}{colors to use (one per biological condition)}

\item{outfile}{TRUE to export the figure in a png file}

\item{ggplot_theme}{ggplot2 theme function (\code{theme_gray()} by default)}
}
\value{
A file named MDS.png in the figures directory
}
\description{
Multi-Dimensional Scaling plot of samples based on the 500 most variant features (for edgeR analyses)
}
\author{
Marie-Agnes Dillies and Hugo Varet
}
