% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeResults.DESeq2.r
\name{summarizeResults.DESeq2}
\alias{summarizeResults.DESeq2}
\title{Summarize DESeq2 analysis}
\usage{
summarizeResults.DESeq2(out.DESeq2, group, independentFiltering = TRUE,
  cooksCutoff = TRUE, alpha = 0.05, col = c("lightblue", "orange",
  "MediumVioletRed", "SpringGreen"), log2FClim = NULL, padjlim = NULL)
}
\arguments{
\item{out.DESeq2}{the result of \code{run.DESeq2()}}

\item{group}{factor vector of the condition from which each sample belongs}

\item{independentFiltering}{\code{TRUE} or \code{FALSE} to perform the independent filtering or not}

\item{cooksCutoff}{outliers detection threshold (TRUE to let DESeq2 choosing it or FALSE to disable the outliers detection)}

\item{alpha}{significance threshold to apply to the adjusted p-values}

\item{col}{colors for the plots}

\item{log2FClim}{numeric vector containing both upper and lower y-axis limits for all the MA-plots produced (NULL by default to set them automatically)}

\item{padjlim}{numeric value between 0 and 1 for the adjusted p-value upper limits for all the volcano plots produced (NULL by default to set them automatically)}
}
\value{
A list containing: (i) a list of \code{data.frames} from \code{exportResults.DESeq2()}, (ii) the table summarizing the independent filtering procedure and (iii) a table summarizing the number of differentially expressed features
}
\description{
Summarize DESeq2 analysis: diagnotic plots, dispersions plot, summary of the independent filtering, export results...
}
\author{
Hugo Varet
}
