% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabSERE}
\alias{tabSERE}
\title{SERE statistics for several samples}
\usage{
tabSERE(counts)
}
\arguments{
\item{counts}{\code{matrix} of raw counts}
}
\value{
The \code{matrix} of SERE values
}
\description{
Compute the SERE statistic for each pair of samples
}
\author{
Marie-Agnes Dillies and Hugo Varet
}

